Chat commands + OptionManager refactoring
