+++ src/game/scripts/vscripts/abilities/hero_perks/npc_dota_hero_skeleton_king_perk.lua
@@ -1,5 +1,4 @@
 local Timers = require('easytimers')
-local Options = require('optionmanager')
 --------------------------------------------------------------------------------------------------------
 --
 --      Hero: Wraith King
