+++ src/game/scripts/vscripts/pregame.lua
@@ -25,7 +25,6 @@ end
 -- Libraries
 local constants = require('constants')
 local network = require('network')
-local OptionManager = require('optionmanager')
 local SkillManager = require('skillmanager')
 local SU = require('lib/StatUploaderFunctions')
 local Timers = require('easytimers')
