+++ src/game/scripts/vscripts/skillmanager.lua
@@ -3,7 +3,6 @@
 ]]
 
 -- Load requires
-local OptionManager = require('optionmanager')
 local util = require('util')
 local Network = require('network')
 local constants = require('constants')
